Fixed some UI tracebacks and code
Regarding the plugin manager being closed before loading completely. Had to check widget.exists() before performing any action.

And about the spinners, I made a function to safely control it because it was also tracebacking if you delete it before it tries to hide itself.

And for the code, I cleaned some indents.